Rename fragments and related classes
