+++ code/app/src/main/java/ca/ualberta/compileorcry/ui/add/AddFragment.java
@@ -1,4 +1,4 @@
-package ca.ualberta.compileorcry.ui.dashboard;
+package ca.ualberta.compileorcry.ui.add;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -12,20 +12,20 @@ import androidx.lifecycle.ViewModelProvider;
 
 import ca.ualberta.compileorcry.databinding.FragmentDashboardBinding;
 
-public class DashboardFragment extends Fragment {
+public class AddFragment extends Fragment {
 
     private FragmentDashboardBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        DashboardViewModel dashboardViewModel =
-                new ViewModelProvider(this).get(DashboardViewModel.class);
+        AddViewModel addViewModel =
+                new ViewModelProvider(this).get(AddViewModel.class);
 
         binding = FragmentDashboardBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
         final TextView textView = binding.textDashboard;
-        dashboardViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        addViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
