+++ code/app/src/main/java/ca/ualberta/compileorcry/ui/feed/FeedViewModel.java
@@ -1,16 +1,16 @@
-package ca.ualberta.compileorcry.ui.dashboard;
+package ca.ualberta.compileorcry.ui.feed;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class DashboardViewModel extends ViewModel {
+public class FeedViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public DashboardViewModel() {
+    public FeedViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is dashboard fragment");
+        mText.setValue("This is home fragment");
     }
 
     public LiveData<String> getText() {
